Language.Haskell.Parser

Description

Haskell parser.

Documentation

parseModule :: String -> ParseResult HsModule

Parse of a string, which should contain a complete Haskell 98 module.

parseModuleWithMode :: ParseMode -> String -> ParseResult HsModule

Parse of a string, which should contain a complete Haskell 98 module.

data ParseMode

Static parameters governing a parse. More to come later, e.g. literate mode, language extensions.

Constructors

ParseMode

Fields

parseFilename :: String

original name of the file being parsed

defaultParseMode :: ParseMode

Default parameters for a parse, currently just a marker for an unknown filename.

data ParseResult a

The result of a parse.

Constructors

ParseOk a
The parse succeeded, yielding a value.

ParseFailed SrcLoc String
The parse failed at the specified source location, with an error message.